requirements.txt
pytest
rapidfuzz
pytesseract
discord.py
Pillow
pytest-asyncio
python-dotenv

